Refactoring: String concatenation with StringBuilder
type: Performance; reason: The reason for using StringBuilder instead of the string concatenation operator is that the concatenation operator creates a new string object each time it is used. This can result in a large number of objects being created in memory, which can impact performance and memory usage. In contrast, the StringBuilder class creates a single buffer that can be appended to, resulting in fewer objects being created.