Template Pattern
the GraphSearchAlgorithm is an abstract class that defines the basic structure of the search algorithm using a graph. It has an abstract method addNodeToQueue and addNeighborsToQueue, which subclasses BFS and DFS implement to provide their specific implementations.  The search method in the GraphSearchAlgorithm class provides the overall structure of the search algorithm using the template method pattern. It initializes a queue and a set of visited nodes and adds the start node to the queue. It then enters a loop that dequeues the next path, checks if the last node in the path has already been visited, adds it to the visited nodes set, and checks if it is the destination node. If it is, it returns the path. Otherwise, it adds the neighbors of the current node to the queue using the addNeighborsToQueue method, if they have not been visited before.  The BFS and DFS classes extend the GraphSearchAlgorithm class and implement the addNodeToQueue and addNeighborsToQueue methods based on the specific requirements of each search algorithm. In particular, BFS adds nodes to the end of the queue, while DFS adds them to the front of the queue. Also, BFS uses a FIFO queue to traverse the graph while DFS uses a LIFO stack.